Create CalcRelease2.c
This is the second release of my milestone project. Key additions to this version is the array that will give each of the functions a function ID depending on which operator is being used in the function. The other key addition is the user defined function that will introduce the program, state the creator of the program, and prompt the user to get started.